PR 88: Add better support for using a local SQL database
Add better support for using a local SQL database.

Note that this still doesn't cover the missing stored procedures, but it does allow the use of a local SQL server instead of Sqlite. The next step will be getting off the stored procedures which will be easier by using a local SQL server.

Ideally, we'd get off Sqlite completely, but the integration tests rely on it still since there won't be a SQL server to hit from the buddy deployments.