PR 95: Add OAuth support for APIs
This change adds OAuth support using the server itself as the connect server. This allows a standard authentication solution for the APIs which doesn't rely on browser cookies.

For context, I'm writing a script which will import user data from prod to a dev machine. Getting and using the auth cookie is near impossible from the script, so adding OAuth makes it easier.

Related work items: #135